requirements.txt.txt
streamlit
pandas
google-genai
plotly
numpy
